Implement a hold-to-confirm screen when exiting game using escape key
